fix: align Windows process status types
Compare the Windows process exit code against STILL_ACTIVE using a matching u32 type so conductor-server compiles on the windows-latest runner.

Validation:
- cargo clippy --workspace -- -D warnings
- cargo test -p conductor-server state::tmux_runtime::tests::restore_runtime_sessions_reattaches_running_tmux_sessions -- --nocapture

+++ crates/conductor-server/src/state/workspace.rs
@@ -490,7 +490,7 @@ pub(crate) fn is_process_alive(pid: u32) -> bool {
             let mut exit_code = 0u32;
             let result = GetExitCodeProcess(handle, &mut exit_code);
             CloseHandle(handle);
-            result != 0 && exit_code == STILL_ACTIVE
+            result != 0 && exit_code == STILL_ACTIVE as u32
         }
     }
